Modified all DB calls to be async in BaseRepository and Controllers (also modified tests)

+++ Project/VetAppointmentAPI/VetAppointmentAPI/VetAppointment.WebAPI/Controllers/OfficesController.cs
@@ -23,16 +23,16 @@ namespace VetAppointment.WebAPI.Controllers
 
         // GET: api/<OfficesController>
         [HttpGet]
-        public IActionResult Get()
+        public async Task<IActionResult> Get()
         {
-            return Ok(officeRepository.All());
+            return Ok(await officeRepository.All());
         }
 
         // GET api/<OfficesController>/5
         [HttpGet("{id}")]
-        public IActionResult Get(Guid id)
+        public async Task<IActionResult> Get(Guid id)
         {
-            var office = officeRepository.Get(id);
+            var office = await officeRepository.Get(id);
             if(office == null)
             {
                 return NotFound();
@@ -42,42 +42,44 @@ namespace VetAppointment.WebAPI.Controllers
 
         // POST api/<OfficesController>
         [HttpPost]
-        public IActionResult Post([FromBody] OfficeDto officeDto)
+        public async Task<IActionResult> Post([FromBody] OfficeDto officeDto)
         {
             var validation = officeValidator.Validate(officeDto);
             if (!validation.IsValid)
                 return StatusCode(400, validation.Errors.First().ErrorMessage);
             Office office = new Office(officeDto.Address);
-            officeRepository.Add(office);
-            officeRepository.SaveChanges();
+            await officeRepository.Add(office);
+            await officeRepository.SaveChanges();
 
             return Created(nameof(Office), office);
         }
 
         // PUT api/<OfficesController>/5
         [HttpPut]
-        public IActionResult Put([FromBody] OfficeDto officeDto)
+        public async Task<IActionResult> Put([FromBody] OfficeDto officeDto)
         {
             var validation = officeValidator.Validate(officeDto);
             if (!validation.IsValid)
                 return StatusCode(400, validation.Errors.First().ErrorMessage);
-            Office? office = officeRepository.Get(officeDto.OfficeId);
+            Office? office = await officeRepository.Get(officeDto.OfficeId);
             if (office == null)
                 return NotFound();
 
-            officeRepository.Update(office);
-            officeRepository.SaveChanges();
+            await officeRepository.Update(office);
+            await officeRepository.SaveChanges();
 
             return NoContent();
         }
 
         // DELETE api/<OfficesController>/5
         [HttpDelete("{id}")]
-        public IActionResult Delete(Guid id)
+        public async Task<IActionResult> Delete(Guid id)
         {
-            Office office = officeRepository.Get(id);
-            officeRepository.Delete(office);
-            officeRepository.SaveChanges();
+            var office = await officeRepository.Get(id);
+            if (office == null)
+                return NotFound();
+            await officeRepository.Delete(office);
+            await officeRepository.SaveChanges();
 
             return NoContent();
         }
